Refactor command buffer tests
Some tests were incorrectly registered under both
`test_cl_khr_comand_buffer` and `test_cl_khr_mutable_dispatch` because
the registration macro was used in a source file that was included in
both executables.

Split the test helper struct definitions and test registration into
separate source files, so the tests are only registered under
`test_cl_khr_command_buffer` which matches the existing behaviour.
